requirements.txt
PyJWT==2.0.1
aiohttp==3.7.3
urllib3==1.26.2
requests==2.25.1
Python==3.8.6